requirements.txt
rapidocr
onnxruntime
pytesseract
Pillow
ollama
pydantic
